Add something1.

+++ LPLD_Project_Template/Template/app/_cal/servoPID.h
@@ -7,9 +7,9 @@
 		float Kp;
 		float Ti;
 		float Td;
-		int32 nowe;		//e(k)
-		int32 laste;	//e(k-1)
-		int32 paste;	//e(k-2)
+		float nowe;		//e(k)
+		float laste;	//e(k-1)
+		float paste;	//e(k-2)
 		int32 dout;
 		int32 out;
 	} ServoPIDTypeDef;
